refactor(input): extract input handlers into dedicated modules
- created app module for application state and input with separated handlers
- moved liststate into appstate
- simplified event loop with action enum
