+++ src/app/action.rs
@@ -1,0 +1,8 @@
+#[derive(Debug, Clone, PartialEq, Default)]
+pub enum Action {
+    #[default]
+    None,
+    #[allow(dead_code)]
+    Exit,
+    Execute(String),
+}
